Separate keypair name from filename
Currently, the name of the AWS or OpenStack keypair must match the
filename of the corresponding private key. AWS, by default, uses a
".pem" extension on private keys generated for download. With the
existing restriction, such private keys must be renamed to work with
the deployment script.

This change decouples the name of the private key from the name of the
keypair used in AWS or OpenStack.  Now, instead of KEYPAIR_PATH (which
pointed to the directory), we use KEYPAIR_FILE to point to the actual
private key file of the keypair being used to deploy.  This allows the
keypair and the associated local private key file to be named
independently.