Install MTA for ActionMailer :sendmail option
We install Postfix as a MTA to provide a Sendmail interface for
ActionMailer to use.  The MTA is configured to listen only on
localhost.  This means it is not accessible for the purposes of
incoming mail from the wider Internet.  This will allow outgoing mail
but prevent possible abuse from hosts attempting to send/relay incoming
e-mail.