Install Solr configuration files from application
The post-1.0 version of the geoblacklight gem generates applications
that include the Solr configuration files as part of the generated
application files.  (These are intended to be used by solr_wrapper.)
Accordingly, we use these files instead of fetching them from
geoblacklight-schema on GitHub, which may be out of sync.

Note that the post-1.0 geoblacklight gem is geared towards Solr 6
compatibility.

Note also that we specify the "geoblacklight_1.1.0" branch of our
application because this has the necessary Solr configuration files
that other branches currently lack.