Fix fallout from geoblacklight gem update to v1.1.2
Version 1.1.2 of the geoblacklight gem brings a requirement of RubyGems
v2.5.2 or later.  The version bundled with the current stable version
of Ruby (2.3.1) is only 2.5.1 so we have to update RubyGems itself via
"gem update --system".  (We also take this opportunity to upgrade to
Ruby 2.3.)

Updating RubyGems appears to cause some knock-on problems with building
the bcrypt gem as an unprivileged user.  Thus, we now perform "bundle
install" as root.